pyproject.toml
[tool.poetry]
name = "econokey-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.0"
flask-restx = "^0.5"
flask-marshmallow = "^0.14"
cryptography = "^37.0.2"
bcrypt = "^3.2.2"
Flask-JWT-Extended = "^4.4.1"
flask-sqlalchemy = ""


[tool.poetry.dev-dependencies]
pylint = "^2"
black = "^"
flake8 = "^4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
